+++ joljak-application/src/main/java/kr/joljak/api/filter/JwtFilter.java
@@ -8,9 +8,9 @@ import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import kr.joljak.jwt.InvalidTokenException;
-import kr.joljak.jwt.JwtTokenProvider;
-import kr.joljak.jwt.PermissionException;
+import kr.joljak.core.jwt.InvalidTokenException;
+import kr.joljak.core.jwt.JwtTokenProvider;
+import kr.joljak.core.jwt.PermissionException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
