+++ joljak-application/src/main/java/kr/joljak/api/invite/controller/InviteController.java
@@ -1,8 +1,8 @@
 package kr.joljak.api.invite.controller;
 
 import io.swagger.annotations.ApiOperation;
-import kr.joljak.invite.entity.Invite;
-import kr.joljak.invite.service.InviteService;
+import kr.joljak.domain.invite.entity.Invite;
+import kr.joljak.domain.invite.service.InviteService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.CrossOrigin;
