+++ joljak-application/src/main/java/kr/joljak/api/user/request/SignUpRequest.java
@@ -3,9 +3,9 @@ package kr.joljak.api.user.request;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import kr.joljak.security.UserRole;
-import kr.joljak.user.entity.User;
-import kr.joljak.user.entity.UserProjectRole;
+import kr.joljak.core.security.UserRole;
+import kr.joljak.domain.user.entity.User;
+import kr.joljak.domain.user.entity.UserProjectRole;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
