+++ joljak-core/src/main/java/kr/joljak/core/jwt/AccessToken.java
@@ -1,0 +1,13 @@
+package kr.joljak.core.jwt;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class AccessToken {
+  private String token;
+  private long expire;
+}
