+++ joljak-core/src/main/java/kr/joljak/core/jwt/InvalidTokenException.java
@@ -1,0 +1,15 @@
+package kr.joljak.core.jwt;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.server.ResponseStatusException;
+
+public class InvalidTokenException extends ResponseStatusException {
+
+  public InvalidTokenException() {
+    this("invalid token exception");
+  }
+
+  public InvalidTokenException(String message) {
+    super(HttpStatus.UNAUTHORIZED, message);
+  }
+}
