+++ joljak-core/src/main/java/kr/joljak/core/jwt/JwtToken.java
@@ -1,0 +1,16 @@
+package kr.joljak.core.jwt;
+
+public enum JwtToken {
+  ACCESS_TOKEN("AccessToken"),
+  REFRESH_TOKEN("RefreshToken");
+
+  private String name;
+
+  JwtToken(String name) {
+    this.name = name;
+  }
+
+  public String getName() {
+    return name;
+  }
+}
