+++ joljak-core/src/main/java/kr/joljak/core/jwt/PermissionException.java
@@ -1,0 +1,14 @@
+package kr.joljak.core.jwt;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.server.ResponseStatusException;
+
+public class PermissionException extends ResponseStatusException {
+  public PermissionException() {
+    this("you don't have permission.");
+  }
+
+  public PermissionException(String message) {
+    super(HttpStatus.FORBIDDEN, message);
+  }
+}
