+++ joljak-core/src/main/java/kr/joljak/core/security/UserRole.java
@@ -1,0 +1,20 @@
+package kr.joljak.core.security;
+
+public enum UserRole {
+  ADMIN("ROLE_ADMIN"),
+  USER("ROLE_USER");
+
+  private String roleName;
+
+  UserRole(String roleName) {
+    this.roleName = roleName;
+  }
+
+  public String getKey() {
+    return name();
+  }
+
+  public String getRoleName() {
+    return roleName;
+  }
+}
