--- joljak-core/src/main/java/kr/joljak/jwt/AccessToken.java
@@ -1,13 +1,0 @@
-package kr.joljak.jwt;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class AccessToken {
-  private String token;
-  private long expire;
-}
