--- joljak-core/src/main/java/kr/joljak/jwt/InvalidTokenException.java
@@ -1,15 +1,0 @@
-package kr.joljak.jwt;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.server.ResponseStatusException;
-
-public class InvalidTokenException extends ResponseStatusException {
-
-  public InvalidTokenException() {
-    this("invalid token exception");
-  }
-
-  public InvalidTokenException(String message) {
-    super(HttpStatus.UNAUTHORIZED, message);
-  }
-}
