--- joljak-core/src/main/java/kr/joljak/jwt/JwtToken.java
@@ -1,16 +1,0 @@
-package kr.joljak.jwt;
-
-public enum JwtToken {
-  ACCESS_TOKEN("AccessToken"),
-  REFRESH_TOKEN("RefreshToken");
-
-  private String name;
-
-  JwtToken(String name) {
-    this.name = name;
-  }
-
-  public String getName() {
-    return name;
-  }
-}
