--- joljak-domain-rds/src/main/java/kr/joljak/common/entity/ExtendEntity.java
@@ -1,22 +1,0 @@
-package kr.joljak.common.entity;
-
-import java.time.LocalDateTime;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class ExtendEntity {
-
-  @CreatedDate
-  private LocalDateTime createdDate;
-
-  @LastModifiedDate
-  private LocalDateTime modifiedDate;
-
-}
