+++ joljak-domain-rds/src/main/java/kr/joljak/domain/user/entity/UserProjectRole.java
@@ -1,0 +1,22 @@
+package kr.joljak.domain.user.entity;
+
+public enum UserProjectRole {
+  DEVELOPER("개발자"),
+  DESIGNER("디자이너"),
+  MEDIA_ART("미디어아트"),
+  PLANNER("기획자");
+
+  private String name;
+
+  UserProjectRole(String name) {
+    this.name = name;
+  }
+
+  public String getKey() {
+    return name();
+  }
+
+  public String getRoleName() {
+    return name;
+  }
+}
