+++ joljak-domain-rds/src/main/java/kr/joljak/domain/user/exception/AlreadyClassOfExistException.java
@@ -1,0 +1,16 @@
+package kr.joljak.domain.user.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.server.ResponseStatusException;
+
+public class AlreadyClassOfExistException extends ResponseStatusException {
+
+  public AlreadyClassOfExistException() {
+    this("already classOf is exist.");
+  }
+
+  public AlreadyClassOfExistException(String message) {
+    super(HttpStatus.CONFLICT, message);
+  }
+}
+
