--- joljak-domain-rds/src/main/java/kr/joljak/invite/exception/AlreadyExistInviteException.java
@@ -1,15 +1,0 @@
-package kr.joljak.invite.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.server.ResponseStatusException;
-
-public class AlreadyExistInviteException extends ResponseStatusException {
-
-  public AlreadyExistInviteException() {
-    this("already exist invite.");
-  }
-
-  public AlreadyExistInviteException(String message) {
-    super(HttpStatus.CONFLICT, message);
-  }
-}
