--- joljak-domain-rds/src/main/java/kr/joljak/invite/exception/InvalidInviteException.java
@@ -1,15 +1,0 @@
-package kr.joljak.invite.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.server.ResponseStatusException;
-
-public class InvalidInviteException extends ResponseStatusException {
-
-  public InvalidInviteException() {
-    this("invalid invite exception");
-  }
-
-  public InvalidInviteException(String message) {
-    super(HttpStatus.UNAUTHORIZED, message);
-  }
-}
