--- joljak-domain-rds/src/main/java/kr/joljak/user/exception/AlreadyClassOfExistException.java
@@ -1,16 +1,0 @@
-package kr.joljak.user.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.server.ResponseStatusException;
-
-public class AlreadyClassOfExistException extends ResponseStatusException {
-
-  public AlreadyClassOfExistException() {
-    this("already classOf is exist.");
-  }
-
-  public AlreadyClassOfExistException(String message) {
-    super(HttpStatus.CONFLICT, message);
-  }
-}
-
